Sort scoreboard
